requirements.txt
asgiref==3.10.0
Django==5.2.7
django-ckeditor==6.7.3
django-jazzmin==3.0.1
django-jet==1.0.8
django-js-asset==3.1.2
django-mptt==0.18.0
django-multiselectfield==1.0.1
Faker==37.11.0
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
